+++ db/zentao.sql
@@ -1295,7 +1295,7 @@ CREATE INDEX `obejctID`   ON `zt_log`(`objectID`);
 
 -- DROP TABLE IF EXISTS `zt_errorlog`;
 CREATE TABLE IF NOT EXISTS `zt_errorlog` (
-  `id` int unsigned NOT NULL AUTO_INCREMENT COMMENT '主键ID',
+  `id` bigint unsigned NOT NULL AUTO_INCREMENT COMMENT '主键ID',
   `requestID` varchar(64) NOT NULL DEFAULT '' COMMENT '请求ID',
   `account` varchar(30) NOT NULL DEFAULT '' COMMENT '用户账号',
   `module` varchar(30) NOT NULL DEFAULT '' COMMENT '模块',
